Refactor quarto-api with dependency inversion and registry pattern
Eliminate circular dependencies by restructuring quarto-api into modular
src/core/api/ architecture with dependency inversion. Move shared utilities from
execute/ to core/, refactor into 13 modules with registry pattern for eager
validation and lazy initialization, and replace Proxy-based API with explicit
getQuartoAPI() function. Also implement dynamic engine registration to resolve
module initialization order issues.
