+++ src/core/pandoc/pandoc-partition.ts
@@ -109,3 +109,24 @@ export function markdownWithExtractedHeading(markdown: string) {
     contentBeforeHeading,
   };
 }
+
+export function languagesInMarkdownFile(file: string) {
+  return languagesInMarkdown(Deno.readTextFileSync(file));
+}
+
+export function languagesInMarkdown(markdown: string) {
+  // see if there are any code chunks in the file
+  const languages = new Set<string>();
+  const kChunkRegex = /^[\t >]*```+\s*\{([a-zA-Z0-9_]+)( *[ ,].*)?\}\s*$/gm;
+  kChunkRegex.lastIndex = 0;
+  let match = kChunkRegex.exec(markdown);
+  while (match) {
+    const language = match[1].toLowerCase();
+    if (!languages.has(language)) {
+      languages.add(language);
+    }
+    match = kChunkRegex.exec(markdown);
+  }
+  kChunkRegex.lastIndex = 0;
+  return languages;
+}
